+++ src/main/java/com/shaadi/repository/UserRepository.java
@@ -7,6 +7,6 @@ import com.shaadi.entity.User;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer>, JpaSpecificationExecutor<User> {
+public interface UserRepository extends JpaRepository<User, Long>, JpaSpecificationExecutor<User> {
     Optional<User> findByEmail(String email);
 }
